fourier: split fourier_sincos and rename
